Changed rudder and sail angles to double and added a start to calculating rudder angle command.

+++ src/main/java/fi/robosailboat/webservice/boatCommunication/dto/Command.java
@@ -5,10 +5,10 @@ import lombok.Data;
 @Data
 public class Command {
 
-    private int r; // rudder angle
-    private int s; // sail angle
+    private double r; // rudder angle
+    private double s; // sail angle
 
-    public Command(final int rudderAngle, final int sailAngle){
+    public Command(final double rudderAngle, final double sailAngle){
         this.r = rudderAngle;
         this.s = sailAngle;
     }
